requirements.txt
Flask==3.0.0
python-vlc==3.0.20123
Werkzeug==3.0.1
flask_socketio=5.5.1